Make UploadService progress notifications arrive more consistently
We were seeing dramatically increased incidence of upload timeouts
which seems to be because the onProgress handler inside the RNBU
Module was not getting fired.

According to extensive documentation here:
https://github.com/gotev/android-upload-service/wiki/Monitoring-upload-status
the UploadStatusDelegate pattern is potentially unreliable. I cannot
actually confirm that we were running into that exact problem. What I
can confirm is that we have seen progress notifications go missing
since I've been working on our app. It seems to be fairly random, but
it does happen, and when it does, the upload will always eventually
get cancelled (by our application) because we believe the upload has
died.

I've made changes to register a single listener for all uploads that
simply sends the notification back through the same RN bridge channel
as before.

Anecdotally, this seems to make those notifications reliable. However,
since this was not always reproducible, the only way we'll know for
sure is to start testing. :/

If this does work, we should consider making a PR to RNBU, as that
project seems newly active.

Technically, the ideal approach would be to move off of UploadService
3.x to 4.x, as mentioned in that documentation. However, there would
likely be significant changes necessary there, and we're probably not
in a great place to be making those at this time if we can get this
simpler approach to work.